[DROOLS-2284] Refactor of toTypedExpression
Moved ToTypedExpression in a separated class

Removed RuleContext, PackageModel from all method of ConstraintParser.java

Introduced new method call that instantiate the context

Introduced TypedExpression context

Removed parentExpression

Partial application of  patternType, bindingId, isPositional

Removed all 9 parameters to ExpressionTyper method

Removed type resolver dependency
